Type the two threshold columns as the bigint the DMV declares, and order the README rung table (review nits)
stale_query_threshold_days and max_plans_per_query are bigint in
sys.database_query_store_options, like every other numeric on the row -
they were collected as int, harmless at realistic values but inconsistent
and a latent OverflowException. Safe to change end-to-end (collector
types, V76 rung, Lite golden, both row classes and readers) because no
store has run V76 yet; the generated-vs-rung parity assertion follows the
generator automatically.

Also moved the V76 row in Darling/README.md's notable-rungs table to
version order (it landed between V46 and V47).

+++ Darling/PerformanceMonitor.Darling.Storage/PgMigrations.cs
@@ -1714,8 +1714,8 @@ CREATE TABLE IF NOT EXISTS collect.query_store_health (
     current_storage_size_mb bigint,
     max_storage_size_mb bigint,
     size_based_cleanup_mode text,
-    stale_query_threshold_days integer,
-    max_plans_per_query integer,
+    stale_query_threshold_days bigint,
+    max_plans_per_query bigint,
     interval_length_minutes bigint
 );
